refactor: 연산자 객체 생성 모듈화
