+++ src/main/java/calculator/Application.java
@@ -1,13 +1,13 @@
 package calculator;
 
+import calculator.domain.Formula;
 import calculator.service.Calculator;
 import calculator.ui.FormulaInput;
-import java.util.List;
 
 public class Application {
 
     public static void main(String[] args) {
-        List<String> formula = FormulaInput.getFormula();
+        Formula formula = FormulaInput.getFormula();
         double result = Calculator.run(formula);
         System.out.println(result);
     }
