+++ src/main/java/calculator/service/Calculator.java
@@ -1,5 +1,6 @@
 package calculator.service;
 
+import calculator.domain.Formula;
 import calculator.util.Operation;
 import java.util.List;
 import java.util.Stack;
@@ -9,16 +10,18 @@ public class Calculator {
     private Calculator() {
     }
 
-    public static double run(List<String> formula){
+    public static double run(Formula formula) {
+        List<String> formulalist = formula.getFormula();
         Stack<String> formulaStack = new Stack<>();
 
-        for(int i = formula.size()-1; i>=0; i--){
-            formulaStack.push(formula.get(i));
+        for (int i = formulalist.size() - 1; i >= 0; i--) {
+            formulaStack.push(formulalist.get(i));
         }
 
-        while(formulaStack.size() > 1){
+        while (formulaStack.size() > 1) {
             double prev = Double.parseDouble(formulaStack.pop());
-            Operation operator = Operation.fromString(formulaStack.pop()).get();
+            Operation operator = Operation.fromString(formulaStack.pop())
+                .orElseThrow(() -> new IllegalArgumentException("연산자 형식 오류"));
             double next = Double.parseDouble(formulaStack.pop());
 
             double result = operator.apply(prev, next);
