+++ src/main/java/calculator/util/Operation.java
@@ -45,7 +45,7 @@ public enum Operation {
             toMap(Object::toString, e -> e));
 
     /**
-     fromString은 toString이 반환하는 문자열을 해당 열거 타입 상수로 변환해주는 기능
+     * fromString은 toString이 반환하는 문자열을 해당 열거 타입 상수로 변환해주는 기능
      */
     public static Optional<Operation> fromString(String symbol) {
         return Optional.ofNullable(stringToEnum.get(symbol));
